modifiylibftfunctionfinishparsingtointtabnextsteptranforminlist

+++ push_swap/Libft/ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: ybarbot <ybarbot@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/11/04 17:17:33 by ybarbot           #+#    #+#             */
-/*   Updated: 2023/11/08 13:44:21 by ybarbot          ###   ########.fr       */
+/*   Updated: 2024/01/22 10:12:12 by ybarbot          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 size_t	ft_strlen(const char *str)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (str[i] != '\0')
